Remove some comments and move container components into containers folder.
