+++ client/containers/ArticleDetail.js
